jwt auth

+++ src/main/java/usermanagement/service/user/impl/UserServiceImpl.java
@@ -1,21 +1,35 @@
 package usermanagement.service.user.impl;
 
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
 import jakarta.transaction.Transactional;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
 import org.springframework.dao.OptimisticLockingFailureException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import usermanagement.dto.UserRegistrationRequest;
+import usermanagement.dto.UserResponse;
 import usermanagement.exception.NotFoundException;
 import usermanagement.exception.UnauthorizedException;
 import usermanagement.mapper.UserMapper;
+import usermanagement.model.ERole;
+import usermanagement.model.Role;
 import usermanagement.model.User;
+import usermanagement.repository.RoleRepository;
 import usermanagement.repository.UserRepository;
 import usermanagement.service.user.UserService;
+import usermanagement.util.JwtTokenProvider;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -25,7 +39,13 @@ public class UserServiceImpl implements UserService {
 
     UserRepository userRepository;
 
+    RoleRepository roleRepository;
+
+    PasswordEncoder passwordEncoder;
+
     UserMapper userMapper;
+
+    JwtTokenProvider jwtTokenProvider;
 
     @Override
     public User findByUsername(String username) {
@@ -33,6 +53,63 @@ public class UserServiceImpl implements UserService {
                 .orElseThrow(() -> new NotFoundException("user not found"));
     }
 
+    @Override
+    public UserResponse getUserResponse(UserRegistrationRequest request) {
+        Optional<User> existingUserWithUsername = userRepository.findByUsername(request.getUsername());
+
+        if (existingUserWithUsername.isPresent()) {
+            throw new NotFoundException("Username already exists");
+        }
+
+        Role userRole = roleRepository.findByName(ERole.USER)
+                .orElseThrow(() -> new NotFoundException("Role not found")); // EDIT
+
+        User user = User.builder()
+                .email(request.getEmail())
+                .username(request.getUsername())
+                .password(passwordEncoder.encode(request.getPassword()))
+                .roles(new HashSet<>(List.of(userRole))) // EDIT
+                .build();
+        userRepository.save(user);
+        return userMapper.toUserResponse(user);
+    }
+
+    @Override
+    public Object getUserByJwt(String jwt) {
+        String userName = jwtTokenProvider.getUserNameFromJwtToken(jwt.substring(7));
+
+        User user = userRepository.findByUsername(userName)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+
+        checkOneTimeToken(jwt);
+        return userMapper.toUserResponse(user);
+    }
+
+    private void checkOneTimeToken(String jwt) {
+        String userName = jwtTokenProvider.getUsername(jwt.substring(7));
+
+        User user = userRepository.findByUsername(userName)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+
+        //TODO: Implement the logic to check if the token is one-time
+
+    }
+
+
+    @Override
+    public ResponseEntity<User> createUser(User user) {
+        Role adminRole = roleRepository.findByName(ERole.ADMIN)
+                .orElseThrow(() -> new NotFoundException("Role not found"));
+
+        User newUser = User.builder()
+                .username(user.getUsername())
+                .password(passwordEncoder.encode(user.getPassword()))
+                .email(user.getEmail())
+                .roles(new HashSet<>(List.of(adminRole)))
+                .build();
+        return ResponseEntity.ok(
+                userRepository.save(newUser));
+    }
 
     @Override
     public List<User> findAllUsers() {
@@ -41,10 +118,6 @@ public class UserServiceImpl implements UserService {
         return users;
     }
 
-    @Override
-    public User createUser(User user) {
-        return userRepository.save(user);
-    }
 
     @Override
     @Transactional
@@ -65,7 +138,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long id) {
+        if (id == null) {
+            throw new NotFoundException("User not found");
+        }
         userRepository.deleteById(id);
     }
-
 }
